chore: Separate command and operation tests for duplicate row removal
